+++ lib/src/surveys/models/posthog_display_survey_question.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'posthog_survey_question_type.dart';
+import 'posthog_display_survey_text_content_type.dart';
 
 /// Base class for all survey questions
 @immutable
@@ -8,6 +9,7 @@ abstract class PostHogDisplaySurveyQuestion {
     required this.type,
     required this.question,
     this.description,
+    this.descriptionContentType = PostHogDisplaySurveyTextContentType.text,
     this.optional = false,
     this.buttonText,
   });
@@ -15,6 +17,7 @@ abstract class PostHogDisplaySurveyQuestion {
   final PostHogSurveyQuestionType type;
   final String question;
   final String? description;
+  final PostHogDisplaySurveyTextContentType descriptionContentType;
   final bool optional;
   final String? buttonText;
 }